Add a confirmation dialog for tools not running in a chat session
